AssociateDescription - Reorganized class inheritance to make testing easier for sets - Refactoring
